Add user option for default line comment prefix
Replace hard-coded line comment prefix with
`specman-default-line-comment-starter` in functions that insert
template text, such as headers, as well as in functions for building
comments.

Make functions obsolete, that insert or edit comments using a specific
line comment style as part of their name. Those are fair game for
removal some time in the future.