Replace scope/list motion with parsing functions
Moving up and down parentheses or curlies blocks is best implemented
using Emacs's own parsing functions.

This patch is the beginning of the removal of specman-mode's
`global-scope-index`. As far as I can see, the functionality around it
is a subset of what can be achieved using standard parsing and
sexp-scanning functions.