Replace `specman-scope-index-down-scope` with `specman-down-scope`
The original implementation of `specman-down-scope` was delegated to
`specman-scope-index-down-scope`, indicating that both functions are
meant to return identical results.

Since the implementation of the former was already replaced with an
algorithm based on `syntax-ppss`, we were now able to retire
`specman-scope-index-down-scope` for good and refactor code from where
it was called.