Switch `specman-beg-of-statement` to syntactic parsing
This is mostly part of an ongoing transition to avoid regexp
searches (particularly `re-search-backward`) where possible.

The original implementation wasn't quite abiding to its own doc string
and didn't actually change point. The only use cases of it called the
function as argument to `goto-char`, which seems odd. This behaviour
has been fixed.

The original implementation has other quirks, some arguably useful,
some not. Let's see where *not* offering bug-for-bug compatibility
bites us and change behaviour accordingly. Example: Point is behind
the semicolon following a statement, but still on the same
line. Technically this position is already part of the new statement.